+++ observable.go
@@ -1,14 +1,12 @@
 package rxgo
 
 import (
-	"sync"
-	"time"
-
 	"github.com/reactivex/rxgo/errors"
 	"github.com/reactivex/rxgo/fx"
 	"github.com/reactivex/rxgo/handlers"
 	"github.com/reactivex/rxgo/optional"
 	"github.com/reactivex/rxgo/options"
+	"sync"
 )
 
 // Observable is a basic observable interface
@@ -378,163 +376,6 @@ func (o *observable) Scan(apply fx.Function2) Observable {
 		close(out)
 	}()
 	return &observable{ch: out}
-}
-
-// Defer waits until an observer subscribes to it, and then it generates an Observable.
-func Defer(f func() Observable) Observable {
-	return &observable{
-		ch:                nil,
-		observableFactory: f,
-	}
-}
-
-// From creates a new Observable from an Iterator.
-func From(it Iterator) Observable {
-	source := make(chan interface{})
-	go func() {
-		for {
-			val, err := it.Next()
-			if err != nil {
-				break
-			}
-			source <- val
-		}
-		close(source)
-	}()
-	return &observable{ch: source}
-}
-
-// Error returns an Observable that invokes an Observer's onError method
-// when the Observer subscribes to it.
-func Error(err error) Observable {
-	return &observable{
-		ch:                  nil,
-		errorOnSubscription: err,
-	}
-}
-
-// Empty creates an Observable with no item and terminate immediately.
-func Empty() Observable {
-	source := make(chan interface{})
-	go func() {
-		close(source)
-	}()
-	return &observable{ch: source}
-}
-
-// Interval creates an Observable emitting incremental integers infinitely between
-// each given time interval.
-func Interval(term chan struct{}, interval time.Duration) Observable {
-	source := make(chan interface{})
-	go func(term chan struct{}) {
-		i := 0
-	OuterLoop:
-		for {
-			select {
-			case <-term:
-				break OuterLoop
-			case <-time.After(interval):
-				source <- i
-			}
-			i++
-		}
-		close(source)
-	}(term)
-	return &observable{ch: source}
-}
-
-// Repeat creates an Observable emitting a given item repeatedly
-func Repeat(item interface{}, ntimes ...int) Observable {
-	source := make(chan interface{})
-
-	// this is the infinity case no ntime parameter is given
-	if len(ntimes) == 0 {
-		go func() {
-			for {
-				source <- item
-			}
-		}()
-		return &observable{ch: source}
-	}
-
-	// this repeat the item ntime
-	if len(ntimes) > 0 {
-		count := ntimes[0]
-		if count <= 0 {
-			return Empty()
-		}
-		go func() {
-			for i := 0; i < count; i++ {
-				source <- item
-			}
-			close(source)
-		}()
-		return &observable{ch: source}
-	}
-
-	return Empty()
-}
-
-// Range creates an Observable that emits a particular range of sequential integers.
-func Range(start, end int) Observable {
-	source := make(chan interface{})
-	go func() {
-		i := start
-		for i < end {
-			source <- i
-			i++
-		}
-		close(source)
-	}()
-	return &observable{ch: source}
-}
-
-// Just creates an Observable with the provided item(s).
-func Just(item interface{}, items ...interface{}) Observable {
-	source := make(chan interface{})
-	if len(items) > 0 {
-		items = append([]interface{}{item}, items...)
-	} else {
-		items = []interface{}{item}
-	}
-
-	go func() {
-		for _, item := range items {
-			source <- item
-		}
-		close(source)
-	}()
-
-	return &observable{ch: source}
-}
-
-// Start creates an Observable from one or more directive-like Supplier
-// and emits the result of each operation asynchronously on a new Observable.
-func Start(f fx.Supplier, fs ...fx.Supplier) Observable {
-	if len(fs) > 0 {
-		fs = append([]fx.Supplier{f}, fs...)
-	} else {
-		fs = []fx.Supplier{f}
-	}
-
-	source := make(chan interface{})
-
-	var wg sync.WaitGroup
-	for _, f := range fs {
-		wg.Add(1)
-		go func(f fx.Supplier) {
-			source <- f()
-			wg.Done()
-		}(f)
-	}
-
-	// Wait in another goroutine to not block
-	go func() {
-		wg.Wait()
-		close(source)
-	}()
-
-	return &observable{ch: source}
 }
 
 func (o *observable) Reduce(apply fx.Function2) OptionalSingle {
